+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ Changelog for package layered_hardware_gazebo
 
 0.0.2 (next version)
 --------------------
-* Add pisition/velocity PID operation mode (effort-based position/velocity mode)
+* Add pisition/velocity/posvel PID operation mode (effort-based position/velocity/posvel mode)
 * Add posvel operation mode (position mode with profile velocity)
 * Add passive operation mode
 * Fix velocity operation mode